Correct Bitfield data type

+++ src/Sia.Skynet/UploadResponse.cs
@@ -11,13 +11,13 @@
         public string Skylink { get; set; }
 
         /// <summary>
-        /// The merkleroot
+        /// The hash that is encoded into the Skylink
         /// </summary>
         public string Merkleroot { get; set; }
 
         /// <summary>
-        /// The bitfield
+        /// The bitfield contains a version, an offset and a length in a heavily compressed and optimized format
         /// </summary>
-        public byte Bitfield { get; set; }
+        public ushort Bitfield { get; set; }
     }
 }
